+++ changelog.md
@@ -8,6 +8,7 @@
 - Indicate pixel coordinates of the decoded image under the mouse pointer
 - Add ESAC as source server
 - Add option to playback at high resolution
+- Switch to Java 17
 
 ## JHelioviewer 4.0.2 (2021-06-11)